@@ -5,10 +5,9 @@
 // gestures. You can also use WidgetTester to find child widgets in the widget
 // tree, read text, and verify that the values of widget properties are correct.
 
-import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import 'package:WOLapp/main.dart';
+import 'package:wol_app/main.dart';
 
 void main() {
   testWidgets('Init smoke test', (WidgetTester tester) async {
